+++ custom_components/ge_home/devices/wac.py
@@ -2,11 +2,12 @@ import logging
 from typing import List
 
 from homeassistant.const import EntityCategory
+from homeassistant.components.binary_sensor import BinarySensorDeviceClass
 from homeassistant.helpers.entity import Entity
 from gehomesdk import ErdCode, ErdApplianceType
 
 from .base import ApplianceApi
-from ..entities import GeWacClimate, GeErdSensor, GeErdBinarySensor, GeErdSwitch, ErdOnOffBoolConverter
+from ..entities import GeWacClimate, GeErdSensor, GeErdBinarySensor, GeErdSwitch, GeErdSelect, ErdOnOffBoolConverter, TurboQuietModeOptionsConverter
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -25,10 +26,24 @@ class WacApi(ApplianceApi):
             GeErdSensor(self, ErdCode.AC_FAN_SETTING, icon_override="mdi:fan", entity_category=EntityCategory.DIAGNOSTIC),
             GeErdSensor(self, ErdCode.AC_OPERATION_MODE, entity_category=EntityCategory.DIAGNOSTIC),
             GeErdSwitch(self, ErdCode.AC_POWER_STATUS, bool_converter=ErdOnOffBoolConverter(), icon_on_override="mdi:power-on", icon_off_override="mdi:power-off"),
-            GeErdBinarySensor(self, ErdCode.AC_FILTER_STATUS, device_class_override="problem", entity_category=EntityCategory.DIAGNOSTIC),
-            GeErdSensor(self, ErdCode.WAC_DEMAND_RESPONSE_STATE, entity_category=EntityCategory.DIAGNOSTIC),
-            GeErdSensor(self, ErdCode.WAC_DEMAND_RESPONSE_POWER, uom_override="kW", entity_category=EntityCategory.DIAGNOSTIC),
+            GeErdBinarySensor(self, ErdCode.AC_FILTER_STATUS, device_class_override=BinarySensorDeviceClass.PROBLEM, entity_category=EntityCategory.DIAGNOSTIC),
         ]
+        if self.has_erd_code(ErdCode.AC_TURBO_QUIET_MODE):
+            available_modes = self.try_get_erd_value(ErdCode.AC_AVAILABLE_TURBO_QUIET_MODES)
+            wac_entities.append(
+                GeErdSelect(
+                    self,
+                    ErdCode.AC_TURBO_QUIET_STATUS,
+                    TurboQuietModeOptionsConverter(
+                        has_turbo=available_modes.has_turbo if available_modes else True,
+                        has_quiet=available_modes.has_quiet if available_modes else True,
+                    ),
+                    control_erd_code=ErdCode.AC_TURBO_QUIET_MODE,
+                    icon_override="mdi:fan-speed-2",
+                    entity_category=EntityCategory.CONFIG,
+                )
+            )
+
         entities = base_entities + wac_entities
         return entities