+++ custom_components/ge_home/entities/__init__.py
@@ -11,3 +11,4 @@ from .water_heater import *
 from .opal_ice_maker import *
 from .ccm import *
 from .dehumidifier import *
+from .laundry import *